🐛 (ProvideAppBarActions): fix state bug where actions would be shown in top app bar under certain conditions where they shouldn't be shown
Specifically, when the display's `WindowWidthSizeClass` changes (e.g., when rotating a phone from portrait to landscape or vice-versa), the actions would be shown in the medium/expanded layout, but would persist upon returning to the compact layout since the actions' state gets restored. This has been fixed by ensuring that the state is conditionally updated and can be reset to `null` based on the value of `shouldShowActions`.