♻️ Use nav arg instead of shared VM to pass MutationType to AddRuleScreen (from SelectRuleTypeScreen)
Instead of in a shared ViewModel scoped to the parent nested nav graph, we store the selected `MutationType` in a regular ViewModel scoped to the SelectRuleTypeScreen, then collect the Flow containing the selected value at the NavHost level and use it to navigate to AddRuleScreen and pass the selected `MutationType` as a navigation argument.

This will reduce complexity when it comes to maintaining user input state (from AddRuleScreen) at the ViewModel level, and ensures that the ViewModel gets cleared when the screen is popped off the back stack while remaining within the same nested nav graph.