💄 (ui): Change colour of CategoryHeading text from primary to secondary
The TabRow in RulesScreen uses the primary colour for its text, so using primary as the text colour in CategoryHeading would result in a busy-looking UI due to overuse of this colour. Switching it to secondary (a more neutral tone) ensures that there is only one source of primary focus.